requirements.txt
streamlit
google-genai
pillow
python-dotenv
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client